requirements.txt
openai>=1.0.0
chromadb>=1.0.0
python-dotenv>=1.0.0
rich>=13.7.0
questionary>=2.0.0
sentence-transformers>=3.0.0
